+++ test/cache/CacheFixture.dart
@@ -1,6 +1,6 @@
 import 'dart:async';
 import 'package:test/test.dart';
-import '../../lib/pip_services3_components.dart';
+import 'package:pip_services3_components/pip_services3_components.dart';
 
 String KEY1 = 'key1';
 String KEY2 = 'key2';