+++ test/lock/LockFixture.dart
@@ -1,5 +1,5 @@
 import 'package:test/test.dart';
-import '../../lib/pip_services3_components.dart';
+import 'package:pip_services3_components/pip_services3_components.dart';
 
 final String LOCK1 = 'lock_1';
 final String LOCK2 = 'lock_2';
@@ -14,73 +14,52 @@ class LockFixture {
 
   void testTryAcquireLock() async {
     // Try to acquire lock for the first time
-    try {
-      var result = await _lock.tryAcquireLock(null, LOCK1, 3000);
-      expect(result, isTrue);
-    } catch (err) {
-      expect(err, isNull);
-    }
+
+    var result = await _lock.tryAcquireLock(null, LOCK1, 3000);
+    expect(result, isTrue);
 
     // Try to acquire lock for the second time
-    try {
-      var result = await _lock.tryAcquireLock(null, LOCK1, 3000);
-      expect(result, isFalse);
-    } catch (err) {
-      expect(err, isNull);
-    }
+
+    result = await _lock.tryAcquireLock(null, LOCK1, 3000);
+    expect(result, isFalse);
 
     // Release the lock
     await _lock.releaseLock(null, LOCK1);
 
     // Try to acquire lock for the third time
-    try {
-      var result = await _lock.tryAcquireLock(null, LOCK1, 3000);
-      expect(result, isTrue);
-    } catch (err) {
-      expect(err, isNull);
-    }
+    result = await _lock.tryAcquireLock(null, LOCK1, 3000);
+    expect(result, isTrue);
 
     await _lock.releaseLock(null, LOCK1);
   }
 
   void testAcquireLock() async {
     // Acquire lock for the first time
-
-    try {
-      await _lock.acquireLock(null, LOCK2, 3000, 1000);
-    } catch (err) {
-      expect(err, isNull);
-    }
+    await _lock.acquireLock(null, LOCK2, 3000, 1000);
 
     // Acquire lock for the second time
+    var err;
     try {
       await _lock.acquireLock(null, LOCK2, 3000, 1000);
-    } catch (err) {
-      expect(err, isNotNull);
+    } catch (ex) {
+      err = ex;
     }
+    expect(err, isNotNull);
 
     // Release the lock
-
     await _lock.releaseLock(null, LOCK2);
 
     // Acquire lock for the third time
-    try {
-      await _lock.acquireLock(null, LOCK2, 3000, 1000);
-    } catch (err) {
-      expect(err, isNull);
-    }
+    await _lock.acquireLock(null, LOCK2, 3000, 1000);
 
     await _lock.releaseLock(null, LOCK2);
   }
 
   void testReleaseLock() async {
     // Acquire lock for the first time
-    try {
-      var result = await _lock.tryAcquireLock(null, LOCK3, 3000);
-      expect(result, isTrue);
-    } catch (err) {
-      expect(err, isNull);
-    }
+
+    var result = await _lock.tryAcquireLock(null, LOCK3, 3000);
+    expect(result, isTrue);
 
     // Release the lock for the first time
     await _lock.releaseLock(null, LOCK3);