+++ test/lock/MemoryLock_test.dart
@@ -1,5 +1,5 @@
 import 'package:test/test.dart';
-import '../../lib/pip_services3_components.dart';
+import 'package:pip_services3_components/pip_services3_components.dart';
 import './LockFixture.dart';
 
 void main() {