+++ test/log/NullLogger_test.dart
@@ -1,7 +1,5 @@
 import 'package:test/test.dart';
-import 'package:pip_services3_commons/pip_services3_commons.dart';
-import '../../lib/src/log/NullLogger.dart';
-import '../../lib/src/log/LogLevel.dart';
+import 'package:pip_services3_components/pip_services3_components.dart';
 
 void main() {
   group('NullLogger', () {
@@ -31,8 +29,8 @@ void main() {
       try {
         // Raise an exception
         throw Exception();
-      } catch (err) {
-        var ex = ApplicationException().wrap(err);
+      } catch (ex) {
+        //var ex = ApplicationException().wrap(err);
         _logger.fatal('123', ex, 'Fatal error');
         _logger.error('123', ex, 'Recoverable error');
         expect(ex, isNotNull);