+++ changelog.md
@@ -13,3 +13,8 @@
 ## 1.0.4
 
 - Fix json converter in LogMessage class.
+
+# 1.0.5 
+
+- Fix acquireLock method in Lock class.
+- Change type of error in ILoger interface methods. Make auto wraping exceptions in ApplicationException type.